Move authentication middleware
