+++ source/B2B.Transactions.Infrastructure/Configuration/CompositionRoot.cs
@@ -20,8 +20,7 @@ using B2B.CimMessageAdapter.Transactions;
 using B2B.Transactions.Authentication;
 using B2B.Transactions.Configuration;
 using B2B.Transactions.DataAccess;
-using B2B.Transactions.Infrastructure.Authentication.Bearer;
-using B2B.Transactions.Infrastructure.Authentication.MarketActors;
+using B2B.Transactions.Infrastructure.Authentication;
 using B2B.Transactions.Infrastructure.DataAccess;
 using B2B.Transactions.Infrastructure.DataAccess.Transaction;
 using B2B.Transactions.Infrastructure.Messages;
